requirements.txt
absl-py==0.11.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
gast==0.3.3
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
six==1.15.0
tensorflow-cpu==2.4.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.12.1
WTForms==2.2.1
